requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
mcp>=1.0.0
